Moved builder code into own subpackage.
Moved builder code into own subpackage.
